Добавил ещё 1 метод для тестирования delete Починил методы тестирования

+++ src/main/java/com/example/gostsNaumen/service/document/DocumentService.java
@@ -67,7 +67,10 @@ public class DocumentService {
     @Transactional
     public void deleteDocumentById(Long id) {
         if (id == null) {
-            throw new EntityNotFoundException("Удаление по пустому ID");
+            throw new IllegalArgumentException("Получен null id");
+        }
+        if (!documentRepository.existsById(id)) {
+            throw new EntityNotFoundException("Документа с таким " + id + " не существует");
         }
         documentRepository.deleteById(id);
     }
